Move fixture class to its own file
